Solve day 7 part 2

+++ java/solutions/src/main/java/com/gusmurphy/fun/aoc/year2025/day7/BeamSplitterSolver.java
@@ -25,7 +25,7 @@ public class BeamSplitterSolver {
                 .sum();
     }
 
-    public static int countPathsIn(String fileName) {
+    public static long countPathsIn(String fileName) {
         var grid = Grid.fromFile(fileName);
         return new PathCounter(grid).countPaths();
     }
